chore(bitgo): refactor tss utils
This change refactors tss utils for reusing common functionalities for different signature schemes.

TICKET: BG-52194
