+++ modules/sdk-core/src/bitgo/utils/tss/baseTypes.ts
@@ -1,0 +1,146 @@
+import { SerializedKeyPair } from 'openpgp';
+import { IRequestTracer } from '../../../api';
+import { KeychainsTriplet } from '../../baseCoin';
+import { Keychain } from '../../keychain';
+import { Memo, WalletType } from '../../wallet/iWallet';
+import { EDDSA } from '../../../account-lib/mpc/tss';
+
+export type TxRequestVersion = 'full' | 'lite';
+
+export interface PrebuildTransactionWithIntentOptions {
+  reqId: IRequestTracer;
+  intentType: string;
+  sequenceId?: string;
+  recipients: {
+    address: string;
+    amount: string | number;
+    tokenName?: string;
+  }[];
+  comment?: string;
+  memo?: Memo;
+  tokenName?: string;
+  nonce?: string;
+}
+
+export type TxRequestState =
+  | 'pendingApproval'
+  | 'canceled'
+  | 'rejected'
+  | 'initialized'
+  | 'pendingDelivery'
+  | 'delivered'
+  | 'pendingUserSignature'
+  | 'signed';
+
+export type TransactionState =
+  | 'initialized'
+  | 'pendingSignature'
+  | 'signed'
+  | 'held'
+  | 'delivered'
+  | 'invalidSignature'
+  | 'rejected';
+
+// Type used to sign a TSS transaction
+export type SignableTransaction = {
+  // unsigned transaction in broadcast format
+  serializedTxHex: string;
+  // portion of a transaction used to generate a signature
+  signableHex: string;
+};
+
+export type UnsignedTransactionTss = SignableTransaction & {
+  // derivation path of the signer
+  derivationPath: string;
+  // transaction fees
+  feeInfo?: {
+    fee: number;
+    feeString: string;
+  };
+  coinSpecific?: Record<string, unknown>;
+  parsedTx?: unknown;
+};
+
+export type TxRequest = {
+  txRequestId: string;
+  walletId: string;
+  walletType: WalletType;
+  version: number;
+  enterpriseId?: string;
+  state: TxRequestState;
+  date: string;
+  userId: string;
+  intent: unknown; // Should override with sig scheme specific intent
+  pendingApprovalId?: string;
+  policiesChecked: boolean;
+  signatureShares?: SignatureShareRecord[];
+  pendingTxHashes?: string[];
+  txHashes?: string[];
+  // Only available in 'lite' version
+  unsignedTxs: UnsignedTransactionTss[]; // Should override with blockchain / sig scheme specific unsigned tx
+  // Only available in 'full' version
+  transactions: {
+    state: TransactionState;
+    unsignedTx: UnsignedTransactionTss; // Should override with blockchain / sig specific unsigned tx
+    signatureShares: SignatureShareRecord[];
+  }[];
+  apiVersion?: TxRequestVersion;
+  latest: boolean;
+};
+
+export enum SignatureShareType {
+  USER = 'user',
+  BACKUP = 'backup',
+  BITGO = 'bitgo',
+}
+
+export interface SignatureShareRecord {
+  from: SignatureShareType;
+  to: SignatureShareType;
+  share: string;
+}
+
+/**
+ * Common Interface for implementing signature scheme specific
+ * util functions
+ */
+export interface ITssUtils<KeyShare = EDDSA.KeyShare> {
+  createUserKeychain(
+    userGpgKey: SerializedKeyPair<string>,
+    userKeyShare: KeyShare,
+    backupKeyShare: KeyShare,
+    bitgoKeychain: Keychain,
+    passphrase: string,
+    originalPasscodeEncryptionCode: string,
+    recipientIndex?: number
+  ): Promise<Keychain>;
+  createBackupKeychain(
+    userGpgKey: SerializedKeyPair<string>,
+    userKeyShare: KeyShare,
+    backupKeyShare: KeyShare,
+    bitgoKeychain: Keychain,
+    passphrase: string
+  ): Promise<Keychain>;
+  createBitgoKeychain(
+    userGpgKey: SerializedKeyPair<string>,
+    userKeyShare: KeyShare,
+    backupKeyShare: KeyShare,
+    enterprise: string
+  ): Promise<Keychain>;
+  createKeychains(params: {
+    passphrase: string;
+    enterprise?: string;
+    originalPasscodeEncryptionCode?: string;
+  }): Promise<KeychainsTriplet>;
+  signTxRequest(params: { txRequest: string | TxRequest; prv: string; reqId: IRequestTracer }): Promise<TxRequest>;
+  prebuildTxWithIntent(
+    params: PrebuildTransactionWithIntentOptions,
+    apiVersion?: TxRequestVersion,
+    preview?: boolean
+  ): Promise<TxRequest>;
+  deleteSignatureShares(txRequestId: string): Promise<SignatureShareRecord[]>;
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  sendTxRequest(txRequestId: string): Promise<any>;
+  recreateTxRequest(txRequestId: string, decryptedPrv: string, reqId: IRequestTracer): Promise<TxRequest>;
+  getTxRequest(txRequestId: string): Promise<TxRequest>;
+}
