+++ modules/sdk-core/src/bitgo/utils/tss/eddsa/types.ts
@@ -1,131 +1,22 @@
-import type { SerializedKeyPair } from 'openpgp';
-import { KeyShare } from '../../../../account-lib/mpc/tss';
-import { IRequestTracer } from '../../../../api';
-import { KeychainsTriplet } from '../../../baseCoin';
-import { Keychain } from '../../../keychain';
-import { Memo } from '../../../wallet';
-import { WalletType } from '../../../wallet/iWallet';
+import { EDDSA } from '../../../../account-lib/mpc/tss';
+import BaseTSSUtils from '../baseTSSUtils';
+import { UnsignedTransactionTss } from '../baseTypes';
 
-export interface PrebuildTransactionWithIntentOptions {
-  reqId: IRequestTracer;
-  intentType: string;
-  sequenceId?: string;
-  recipients: {
-    address: string;
-    amount: string | number;
-    tokenName?: string;
-  }[];
-  comment?: string;
-  memo?: Memo;
-  tokenName?: string;
-  nonce?: string;
-}
+export type KeyShare = EDDSA.KeyShare;
+export type YShare = EDDSA.YShare;
+/** @deprected use UnsignedTransactionTss from baseTypes */
+export type EddsaUnsignedTransaction = UnsignedTransactionTss;
 
-export type TxRequestVersion = 'full' | 'lite';
+export type IEddsaUtils = BaseTSSUtils<KeyShare>;
+/** @deprecated Use IEddsaUtils */
+export type ITssUtils = IEddsaUtils;
 
-export type EddsaUnsignedTransaction = {
-  serializedTxHex: string;
-  signableHex: string;
-  feeInfo?: {
-    fee: number;
-    feeString: string;
-  };
-  derivationPath: string;
-};
-
-export type TxRequestState =
-  | 'pendingApproval'
-  | 'canceled'
-  | 'rejected'
-  | 'initialized'
-  | 'pendingDelivery'
-  | 'delivered'
-  | 'pendingUserSignature'
-  | 'signed';
-
-export type TransactionState =
-  | 'initialized'
-  | 'pendingSignature'
-  | 'signed'
-  | 'held'
-  | 'delivered'
-  | 'invalidSignature'
-  | 'rejected';
-
-export type TxRequest = {
-  txRequestId: string;
-  walletId: string;
-  walletType: WalletType;
-  version: number;
-  enterpriseId?: string;
-  state: TxRequestState;
-  date: string;
-  userId: string;
-  intent: any;
-  pendingApprovalId?: string;
-  policiesChecked: boolean;
-  signatureShares?: SignatureShareRecord[];
-  pendingTxHashes?: string[];
-  txHashes?: string[];
-  // Only available in 'lite' version
-  unsignedTxs: EddsaUnsignedTransaction[];
-  // Only available in 'full' version
-  transactions: {
-    state: TransactionState;
-    unsignedTx: EddsaUnsignedTransaction;
-    signatureShares: SignatureShareRecord[];
-  }[];
-  apiVersion?: TxRequestVersion;
-  latest: boolean;
-};
-
-export enum SignatureShareType {
-  USER = 'user',
-  BACKUP = 'backup',
-  BITGO = 'bitgo',
-}
-
-export interface SignatureShareRecord {
-  from: SignatureShareType;
-  to: SignatureShareType;
-  share: string;
-}
-
-export interface ITssUtils {
-  createUserKeychain(
-    userGpgKey: SerializedKeyPair<string>,
-    userKeyShare: KeyShare,
-    backupKeyShare: KeyShare,
-    bitgoKeychain: Keychain,
-    passphrase: string,
-    originalPasscodeEncryptionCode: string
-  ): Promise<Keychain>;
-  createBackupKeychain(
-    userGpgKey: SerializedKeyPair<string>,
-    userKeyShare: KeyShare,
-    backupKeyShare: KeyShare,
-    bitgoKeychain: Keychain,
-    passphrase: string
-  ): Promise<Keychain>;
-  createBitgoKeychain(
-    userGpgKey: SerializedKeyPair<string>,
-    userKeyShare: KeyShare,
-    backupKeyShare: KeyShare,
-    enterprise: string
-  ): Promise<Keychain>;
-  createKeychains(params: {
-    passphrase: string;
-    enterprise?: string;
-    originalPasscodeEncryptionCode?: string;
-  }): Promise<KeychainsTriplet>;
-  signTxRequest(params: { txRequest: string | TxRequest; prv: string; reqId: IRequestTracer }): Promise<TxRequest>;
-  prebuildTxWithIntent(
-    params: PrebuildTransactionWithIntentOptions,
-    apiVersion?: TxRequestVersion,
-    preview?: boolean
-  ): Promise<TxRequest>;
-  deleteSignatureShares(txRequestId: string): Promise<SignatureShareRecord[]>;
-  sendTxRequest(txRequestId: string): Promise<any>;
-  recreateTxRequest(txRequestId: string, decryptedPrv: string, reqId: IRequestTracer): Promise<TxRequest>;
-  getTxRequest(txRequestId: string): Promise<TxRequest>;
-}
+// For backward compatibility
+export {
+  PrebuildTransactionWithIntentOptions,
+  TxRequestVersion,
+  TxRequestState,
+  TransactionState,
+  SignatureShareRecord,
+  SignatureShareType,
+} from '../baseTypes';
